+++ tests/cert/main.tf
@@ -1,0 +1,55 @@
+variable "conjur_ssl_cert" {}
+variable "conjur_appliance_url" {}
+variable "conjur_account" {}
+variable "conjur_host_id" {}
+variable "conjur_authn_type" {}
+variable "conjur_authn_service_id" {}
+variable "conjur_secret_variable" {}
+variable "conjur_authn_cert" {}
+variable "conjur_authn_cert_key" {}
+
+terraform {
+  required_providers {
+    conjur = {
+      source  = "terraform.example.com/cyberark/conjur"
+      version = "~> 0"
+    }
+  }
+}
+
+provider "conjur" {
+  appliance_url  = var.conjur_appliance_url
+  account        = var.conjur_account
+  authn_type     = var.conjur_authn_type
+  service_id     = var.conjur_authn_service_id
+  host_id        = var.conjur_host_id
+  ssl_cert       = var.conjur_ssl_cert
+  authn_cert     = var.conjur_authn_cert
+  authn_cert_key = var.conjur_authn_cert_key
+}
+
+data "conjur_secret" "dbpass" {
+  name = var.conjur_secret_variable
+}
+
+output "dbpass-to-output" {
+  value     = data.conjur_secret.dbpass.value
+  sensitive = true
+}
+
+ephemeral "conjur_secret" "dbpass" {
+  name = var.conjur_secret_variable
+}
+
+resource "local_file" "dbpass-to-file" {
+  content         = data.conjur_secret.dbpass.value
+  filename        = "${path.module}/../dbpass"
+  file_permission = "0664"
+
+  lifecycle {
+    precondition {
+      condition     = ephemeral.conjur_secret.dbpass.value == data.conjur_secret.dbpass.value
+      error_message = "Ephemeral secret value does not match regular data source value"
+    }
+  }
+}